+++ libs/sdk-core/src/lib/db/resource/resource.ts
@@ -25,8 +25,8 @@ import type {
   UserTokenAnnotation,
 } from './resource.models';
 import { ExtractedDataTypes, ResourceFieldProperties } from './resource.models';
-import type { Chat, ChatOptions, Search, SearchOptions } from '../search';
-import { find, search, chat } from '../search';
+import type { Ask, ChatOptions, Search, SearchOptions } from '../search';
+import { find, search, ask } from '../search';
 import { retry429Config, setEntities, setLabels, sliceUnicode } from './resource.helpers';
 import { RagStrategyName } from '../kb';
 
@@ -381,60 +381,60 @@ export class Resource extends ReadableResource implements IResource {
    * Retrieves a generative answer for the given query based on
    * the results of a search operation performed on the resource.
    */
-  chat(
+  ask(
     query: string,
-    context?: Chat.ContextEntry[],
-    features?: Chat.Features[],
+    context?: Ask.ContextEntry[],
+    features?: Ask.Features[],
     options?: ChatOptions,
-  ): Observable<Chat.Answer | IErrorResponse>;
-  chat(
+  ): Observable<Ask.Answer | IErrorResponse>;
+  ask(
     query: string,
-    context?: Chat.ContextEntry[],
-    features?: Chat.Features[],
+    context?: Ask.ContextEntry[],
+    features?: Ask.Features[],
     options?: ChatOptions,
-    callback?: (answer: Chat.Answer | IErrorResponse) => void,
+    callback?: (answer: Ask.Answer | IErrorResponse) => void,
   ): Observable<null>;
-  chat(
+  ask(
     query: string,
-    context?: Chat.ContextEntry[],
-    features?: Chat.Features[],
+    context?: Ask.ContextEntry[],
+    features?: Ask.Features[],
     options?: ChatOptions,
-    callback?: (answer: Chat.Answer | IErrorResponse) => void,
-  ): Observable<Chat.Answer | IErrorResponse> | Observable<null> {
-    const chatRequest = chat(this.nuclia, this.kb, this.path, query, context, features, options);
+    callback?: (answer: Ask.Answer | IErrorResponse) => void,
+  ): Observable<Ask.Answer | IErrorResponse> | Observable<null> {
+    const askRequest = ask(this.nuclia, this.kb, this.path, query, context, features, options);
     if (callback) {
-      chatRequest.subscribe((response) => callback(response));
+      askRequest.subscribe((response) => callback(response));
       return of(null);
     }
-    return chatRequest;
+    return askRequest;
   }
 
   /**
    * Retrieves a generative answer for the given query using the entire resource as context
    * (the resource's text might be shorten if too large).
    */
-  ask(
+  askToResource(
     query: string,
-    context?: Chat.ContextEntry[],
-    features?: Chat.Features[],
+    context?: Ask.ContextEntry[],
+    features?: Ask.Features[],
     options?: ChatOptions,
-  ): Observable<Chat.Answer | IErrorResponse>;
-  ask(
+  ): Observable<Ask.Answer | IErrorResponse>;
+  askToResource(
     query: string,
-    context?: Chat.ContextEntry[],
-    features?: Chat.Features[],
+    context?: Ask.ContextEntry[],
+    features?: Ask.Features[],
     options?: ChatOptions,
-    callback?: (answer: Chat.Answer | IErrorResponse) => void,
+    callback?: (answer: Ask.Answer | IErrorResponse) => void,
   ): Observable<null>;
-  ask(
+  askToResource(
     query: string,
-    context?: Chat.ContextEntry[],
-    features?: Chat.Features[],
+    context?: Ask.ContextEntry[],
+    features?: Ask.Features[],
     options?: ChatOptions,
-    callback?: (answer: Chat.Answer | IErrorResponse) => void,
-  ): Observable<Chat.Answer | IErrorResponse> | Observable<null> {
+    callback?: (answer: Ask.Answer | IErrorResponse) => void,
+  ): Observable<Ask.Answer | IErrorResponse> | Observable<null> {
     options = { ...(options || {}), rag_strategies: [{ name: RagStrategyName.FULL_RESOURCE }] };
-    return this.chat(query, context, features, options, callback);
+    return this.ask(query, context, features, options, callback);
   }
 
   setLabels(fieldId: string, fieldType: string, paragraphId: string, labels: Classification[]): Observable<void> {