+++ libs/sdk-core/CHANGELOG.md
@@ -1,14 +1,18 @@
-# 1.15.0 (unreleased)
+# 1.15.0 (2024-05-17)
 
 ### Breaking changes
 
-- Remove methods using deprecated `/stats` endpoint and remove corresponding models. 
+- Remove methods using deprecated `/stats` endpoint and remove corresponding models.
+- Rename `Chat` namespace to `Ask`.
+- Rename the existing `.ask()` method in `Resource` to `.askToResource()`
+- Rename `.chat()` method in `KnowledgeBox` and in `Resource` to `.ask()`
 
 ### Improvements
 
 - Update `Citations` model
 - Add `KEEP_MARKDOWN` format to `TextFormat` type
 - Add `creation_date` in Account model
+- Support the new `ask` endpoint
 
 # 1.14.9 (2024-04-30)